# openapi.yaml
openapi: "3.1.0"
info:
  title: "Aegis API"
  version: "1.0.0"
  description: "Orchestrate Claude Code sessions via REST API. Aegis wraps Claude Code in ACP sessions\nand exposes everything through a unified API.\n\n**Base URL:** `http://localhost:9100`\n\n**Authentication:** All endpoints except `/v1/health` and `/health` require a bearer token.\nPass `Authorization: Bearer <your-token>` header, or set `AEGIS_AUTH_TOKEN` env var.\n\n**Session lifecycle:** `working` → `idle` → `permission_prompt` → `asking` → `stalled`\n\n**API versioning:** All `/v1/` responses include `X-Aegis-API-Version: 1`.\nLegacy unversioned paths (e.g. `/health`, `/sessions`) emit `Deprecation: true`,\n`Sunset`, and `X-API-Deprecated` headers. See `docs/api-versioning.md` for the\nfull versioning policy.\n"
servers:
  - url: "http://localhost:9100"
    description: "Local Aegis server"
security:
  - BearerAuth: []
paths:
...
  /v1/auth/verify:
    post:
      operationId: "verifyApiToken"
      summary: "Verify an API token and return its role"
      tags:
        - "Auth"

      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTokenRequest"




      responses:
        "200":
          description: "Token is valid"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyTokenResponse"




        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          description: "Token is invalid"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyTokenResponse"




        "429":
          description: "Token is valid but currently rate limited"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyTokenResponse"
  /v1/auth/keys:
    get:
      operationId: "listApiKeys"
      summary: "List all API keys"
      tags:
        - "Auth"

      responses:
        "200":
          description: "List of API keys (plaintext shown only at creation)"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  keys:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/ApiKey"







        "401":
          $ref: "#/components/responses/Unauthorized"



    post:
      operationId: "createApiKey"
      summary: "Create a new API key"
      tags:
        - "Auth"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: "string"
                  description: "Human-readable name for this key"
                  example: "production-operator"

                expiresAt:
                  type: "string"
                  format: "date-time"
                  description: "Optional expiration timestamp (ISO 8601)"
                  example: "2025-12-31T23:59:59Z"


              required:
                - "name"





      responses:
        "201":
          description: "API key created (plaintext shown only now)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"




        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"
  "/v1/auth/keys/{id}":
    delete:
      operationId: "revokeApiKey"
      summary: "Revoke an API key"
      tags:
        - "Auth"

      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"

          description: "API key ID"


      responses:
        "204":
          description: "API key revoked"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
  /v1/auth/sse-token:
    post:
      operationId: "createSseToken"
      summary: "Generate a short-lived SSE token for browser clients"
      tags:
        - "Auth"

      description: "SSE tokens are short-lived (60s), single-use, and allow EventSource clients\n(which cannot set custom headers) to authenticate SSE subscriptions.\n"
      responses:
        "200":
          description: "SSE token generated"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  token:
                    type: "string"
                    description: "Short-lived SSE token"

                  expiresAt:
                    type: "integer"
                    description: "Unix timestamp of expiration"






        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/sessions:
    get:
      operationId: "listSessions"
      summary: "List all sessions"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/limit"

        - $ref: "#/components/parameters/offset"

        - name: "status"
          in: "query"
          schema:
            $ref: "#/components/schemas/SessionStatusFilter"

          description: "Filter by session UI state"


      responses:
        "200":
          description: "List of sessions"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  sessions:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/SessionInfo"


                  total:
                    type: "integer"






        "401":
          $ref: "#/components/responses/Unauthorized"



    post:
      operationId: "createSession"
      summary: "Create or reuse a session"
      tags:
        - "Sessions"

      description: "Creates a new Claude Code session or reuses an existing idle session for the same `workDir`.\nReturns `201` for new sessions, `200` for reused ones.\n"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"




      responses:
        "200":
          description: "Existing idle session reused"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionReused"




        "201":
          description: "New session created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreated"




        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "409":
          description: "Session creation conflict"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"

                  sessionId:
                    type: "string"
...
  "/v1/auth/keys/{id}/rotate":
    post:
      summary: "Rotate API key"
      operationId: "rotateApiKey"
      tags:
        - "Auth"

      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Key ID"
          schema:
            type: "string"



      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                ttlDays:
                  type: "integer"
                  exclusiveMinimum: 0
                  maximum: 9007199254740991


              additionalProperties: false




      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  id:
                    type: "string"

                  key:
                    type: "string"

                  name:
                    type: "string"

                  expiresAt:
                    anyOf:
                      - type: "number"

                      - type: "null"



                  role:
                    type: "string"
                    enum:
                      - "admin"
                      - "operator"
                      - "viewer"


                  permissions:
                    type: "array"
                    items:
                      type: "string"
                      enum:
                        - "create"
                        - "send"
                        - "approve"
                        - "reject"
                        - "kill"




                required:
                  - "id"
                  - "key"
                  - "name"
                  - "expiresAt"
                  - "role"
                  - "permissions"

                additionalProperties: false




        "404":
          description: "Not found"
...
components:
  securitySchemes:
    BearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "API key"
  parameters:
...
    limit:
      name: "limit"
      in: "query"
      schema:
        type: "integer"
        default: 50
        minimum: 1
        maximum: 200
    offset:
      name: "offset"
      in: "query"
      schema:
        type: "integer"
        default: 0
  responses:
    BadRequest:
      description: "Invalid request body or parameters"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
    Unauthorized:
      description: "Missing or invalid authentication token"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Unauthorized"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Not found"
  schemas:
    ApiKeyRole:
      type: "string"
      enum:
        - "viewer"
        - "operator"
        - "admin"
    VerifyTokenRequest:
      type: "object"
      properties:
        token:
          type: "string"


      required:
        - "token"

      additionalProperties: false
    VerifyTokenResponse:
      type: "object"
      properties:
        valid:
          type: "boolean"

        role:
          $ref: "#/components/schemas/ApiKeyRole"


      required:
        - "valid"
...
    ApiKey:
      type: "object"
      properties:
        id:
          type: "string"

        name:
          type: "string"

        keyId:
          type: "string"
          description: "First 8 characters of the key hash"

        createdAt:
          type: "integer"

        expiresAt:
          type: "integer"
          nullable: true

        lastUsedAt:
          type: "integer"
          nullable: true
    ApiKeyCreated:
      type: "object"
      properties:
        id:
          type: "string"

        name:
          type: "string"

        key:
          type: "string"
          description: "Full plaintext key — shown ONLY here, never again"

        keyId:
          type: "string"

        createdAt:
          type: "integer"
...
    CreateSessionRequest:
      type: "object"
      properties:
        name:
          type: "string"
          description: "Human-readable session name"
          example: "feature-auth"

        workDir:
          type: "string"
          description: "Working directory for Claude Code"
          example: "/home/user/my-project"

        prompt:
          type: "string"
          description: "Initial prompt to send to Claude Code"

        permissionMode:
          type: "string"
          enum:
            - "default"
            - "bypassPermissions"
            - "plan"
            - "acceptEdits"
            - "dontAsk"
            - "auto"

          default: "default"

        autoApprove:
          type: "boolean"
          default: false

        env:
          type: "object"
          additionalProperties:
            type: "string"

          description: "Environment variables to inject into the session"

        claudeCommand:
          type: "string"
          maxLength: 10000
          description: "Custom Claude Code command/flags"

        maxSessionAgeMs:
          type: "integer"
          description: "Maximum session lifetime in milliseconds"

        stallThresholdMs:
          type: "integer"
          description: "Stall detection threshold override"


      required:
        - "workDir"
    SessionCreated:
      type: "object"
      properties:
        id:
          type: "string"

        name:
          type: "string"

        status:
          type: "string"
          example: "working"

        workDir:
          type: "string"

        reused:
          type: "boolean"
          example: false

        promptDelivery:
          type: "object"
          properties:
            delivered:
              type: "boolean"

            attempts:
              type: "integer"
    SessionReused:
      allOf:
        - $ref: "#/components/schemas/SessionCreated"

        - type: "object"
          properties:
            reused:
              type: "boolean"
              example: true
    SessionInfo:
      type: "object"
      properties:
        id:
          type: "string"

        windowId:
          type: "string"

        windowName:
          type: "string"

        workDir:
          type: "string"

        claudeSessionId:
          type: "string"

        jsonlPath:
          type: "string"

        byteOffset:
          type: "integer"

        monitorOffset:
          type: "integer"

        status:
          type: "string"
          enum:
            - "idle"
            - "working"
            - "compacting"
            - "context_warning"
            - "waiting_for_input"
            - "permission_prompt"
            - "plan_mode"
            - "ask_question"
            - "bash_approval"
            - "settings"
            - "error"
            - "rate_limit"
            - "unknown"


        createdAt:
          type: "integer"

        lastActivity:
          type: "integer"

        stallThresholdMs:
          type: "integer"

        permissionStallMs:
          type: "integer"

        permissionMode:
          type: "string"

        settingsPatched:
          type: "boolean"

        hookSettingsFile:
          type: "string"

        hookSecret:
          type: "string"

        lastHookAt:
          type: "integer"

        activeSubagents:
          type: "array"
          items:
            type: "string"


        permissionPromptAt:
          type: "integer"

        permissionRespondedAt:
          type: "integer"

        lastHookReceivedAt:
          type: "integer"

        lastHookEventAt:
          type: "integer"

        model:
          type: "string"

        lastDeadAt:
          type: "integer"

        ccPid:
          type: "integer"

        parentId:
          type: "string"

        children:
          type: "array"
          items:
            type: "string"


        permissionPolicy:
          type: "array"
          items:
            type: "object"


        permissionProfile:
          type: "object"

        prd:
          type: "string"

        ownerKeyId:
          type: "string"
...
    SessionStatusFilter:
      type: "string"
      enum:
        - "all"
        - "idle"
        - "working"
        - "compacting"
        - "context_warning"
        - "waiting_for_input"
        - "permission_prompt"
        - "plan_mode"
        - "ask_question"
        - "bash_approval"
        - "settings"
        - "error"
        - "rate_limit"
        - "unknown"